A default layout and a Minemak
