+++ keyboards/handwired/acacia/info.json
@@ -20,57 +20,57 @@
     "layouts": {
         "LAYOUT": {
             "layout": [
-                {"matrix": [3, 0], "x": 0, "y": 1.5},
-                {"matrix": [0, 0], "x": 1, "y": 1},
-                {"matrix": [0, 1], "x": 2, "y": 0.5},
+                {"matrix": [3, 0], "x": 0, "y": 0.5},
+                {"matrix": [0, 0], "x": 1, "y": 0},
+                {"matrix": [0, 1], "x": 2, "y": 0},
                 {"matrix": [0, 2], "x": 3, "y": 0},
-                {"matrix": [0, 3], "x": 4, "y": 0.5},
-                {"matrix": [3, 3], "x": 5, "y": 1},
+                {"matrix": [0, 3], "x": 4, "y": 0},
+                {"matrix": [3, 3], "x": 5, "y": 0.5},
 
-                {"matrix": [3, 6], "x": 10.5, "y": 1},
-                {"matrix": [0, 6], "x": 11.5, "y": 0.5},
-                {"matrix": [0, 7], "x": 12.5, "y": 0},
-                {"matrix": [0, 8], "x": 13.5, "y": 0.5},
-                {"matrix": [0, 9], "x": 14.5, "y": 1},
-                {"matrix": [3, 9], "x": 15.5, "y": 1.5},
+                {"matrix": [3, 6], "x": 10, "y": 0.5},
+                {"matrix": [0, 6], "x": 11, "y": 0},
+                {"matrix": [0, 7], "x": 12, "y": 0},
+                {"matrix": [0, 8], "x": 13, "y": 0},
+                {"matrix": [0, 9], "x": 14, "y": 0},
+                {"matrix": [3, 9], "x": 15, "y": 0.5},
 
-                {"matrix": [4, 0], "x": 0, "y": 2.5},
-                {"matrix": [1, 0], "x": 1, "y": 2},
-                {"matrix": [1, 1], "x": 2, "y": 1.5},
-                {"matrix": [1, 2], "x": 3, "y": 1.2},
-                {"matrix": [1, 3], "x": 4, "y": 1.5},
-                {"matrix": [4, 3], "x": 5, "y": 2},
+                {"matrix": [4, 0], "x": 0, "y": 1.5},
+                {"matrix": [1, 0], "x": 1, "y": 1},
+                {"matrix": [1, 1], "x": 2, "y": 1},
+                {"matrix": [1, 2], "x": 3, "y": 1},
+                {"matrix": [1, 3], "x": 4, "y": 1},
+                {"matrix": [4, 3], "x": 5, "y": 1.5},
 
-                {"matrix": [4, 6], "x": 10.5, "y": 2},
-                {"matrix": [1, 6], "x": 11.5, "y": 1.5},
-                {"matrix": [1, 7], "x": 12.5, "y": 1.2},
-                {"matrix": [1, 8], "x": 13.5, "y": 1.5},
-                {"matrix": [1, 9], "x": 14.5, "y": 2},
-                {"matrix": [4, 9], "x": 15.5, "y": 2.5},
+                {"matrix": [4, 6], "x": 10, "y": 1.5},
+                {"matrix": [1, 6], "x": 11, "y": 1},
+                {"matrix": [1, 7], "x": 12, "y": 1},
+                {"matrix": [1, 8], "x": 13, "y": 1},
+                {"matrix": [1, 9], "x": 14, "y": 1},
+                {"matrix": [4, 9], "x": 15, "y": 1.5},
 
-                {"matrix": [2, 0], "x": 1, "y": 3},
-                {"matrix": [2, 1], "x": 2, "y": 2.5},
-                {"matrix": [2, 2], "x": 3, "y": 2.2},
-                {"matrix": [2, 3], "x": 4, "y": 2.5},
+                {"matrix": [2, 0], "x": 1, "y": 2},
+                {"matrix": [2, 1], "x": 2, "y": 2},
+                {"matrix": [2, 2], "x": 3, "y": 2},
+                {"matrix": [2, 3], "x": 4, "y": 2},
 
-                {"matrix": [2, 6], "x": 11.5, "y": 2.5},
-                {"matrix": [2, 7], "x": 12.5, "y": 2.2},
-                {"matrix": [2, 8], "x": 13.5, "y": 2.5},
-                {"matrix": [2, 9], "x": 14.5, "y": 3},
+                {"matrix": [2, 6], "x": 11, "y": 2},
+                {"matrix": [2, 7], "x": 12, "y": 2},
+                {"matrix": [2, 8], "x": 13, "y": 2},
+                {"matrix": [2, 9], "x": 14, "y": 2},
 
-                {"matrix": [0, 4], "x": 2.25, "y": 4},
-                {"matrix": [1, 4], "x": 4, "y": 4},
-                {"matrix": [2, 4], "x": 5, "y": 4},
-                {"matrix": [3, 4], "x": 6, "y": 4},
-                {"matrix": [4, 4], "x": 7, "y": 4},
+                {"matrix": [0, 4], "x": 2.25, "y": 3},
+                {"matrix": [1, 4], "x": 3.5, "y": 3},
+                {"matrix": [2, 4], "x": 4.5, "y": 3},
+                {"matrix": [3, 4], "x": 5.5, "y": 3},
+                {"matrix": [4, 4], "x": 6.5, "y": 3},
 
-                {"matrix": [4, 5], "x": 8.5, "y": 4},
-                {"matrix": [3, 5], "x": 9.5, "y": 4},
-                {"matrix": [2, 5], "x": 10.5, "y": 4},
-                {"matrix": [1, 5], "x": 11.5, "y": 4},
-                {"matrix": [0, 5], "x": 13.25, "y": 4},
+                {"matrix": [4, 5], "x": 8.5, "y": 3},
+                {"matrix": [3, 5], "x": 9.5, "y": 3},
+                {"matrix": [2, 5], "x": 10.5, "y": 3},
+                {"matrix": [1, 5], "x": 11.5, "y": 3},
+                {"matrix": [0, 5], "x": 12.75, "y": 3},
 
-                {"matrix": [3, 8], "x": 7.7, "y": 2}
+                {"matrix": [3, 8], "x": 7, "y": 1}
             ]
         }
     }
