+++ keyboards/handwired/acacia/keymaps/default/keymap.c
@@ -15,134 +15,40 @@
  */
 #include QMK_KEYBOARD_H
 
-enum layer_names {
-  _ALPHA,
-  _NAVI,
-  _SYM_NUM,
-  _FUNC
-};
-
-#define MO_SYM MO(_SYM_NUM)
-#define OSM_LSFT OSM(MOD_LSFT)
-#define SPC_CHNG LT(_FUNC, KC_SPC)
-#define LOCK_ALP TO(_ALPHA)
-#define LOCK_NAV TO(_NAVI)
-#define NUL_GUI LGUI_T(KC_0)
+#define SPC_CTL LCTL_T(KC_SPC)
+#define DEL_GUI LGUI_T(KC_DEL)
+#define DEL_CTL LCTL_T(KC_SPC)
+#define OSM_SFT OSM(MOD_LSFT)
+#define OSM_ALT OSM(MOD_LALT)
+#define ESC_CTL LCTL_T(KC_ESC)
+#define BS_GUI LGUI_T(KC_BSPC)
+#define BS_SYM LT(_SYM_NUM, KC_BSPC)
+#define SPC_NAV LT(_NAVI_FN, KC_SPC)
 #define OSM_AGR OSM(MOD_RALT)
-#define Z_GUI LGUI_T(KC_Z)
-#define X_ALT LALT_T(KC_X)
-#define C_SYM LT(_SYM_NUM, KC_C)
-#define V_CTL LCTL_T(KC_V)
-#define G_AGR RALT_T(KC_G)
-#define H_AGR RALT_T(KC_H)
-#define M_CTL RCTL_T(KC_M)
-#define COMM_SYM LT(_SYM_NUM, KC_COMM)
-#define DOT_ALT LALT_T(KC_DOT)
-#define SLSH_GUI RGUI_T(KC_SLSH)
-#define J_GUI RGUI_T(KC_J)
-#define UNDO LCTL(KC_Z)
-#define CUT LCTL(KC_X)
-#define COPY LCTL(KC_C)
-#define PASTE LCTL(KC_V)
-#define QUIT LCTL(KC_Q)
-#define CLOSE LCTL(KC_W)
-#define FIND LCTL(KC_F)
-#define ALL LCTL(KC_A)
-#define SAVE LCTL(KC_S)
-#define MARK LCTL(KC_D)
-#define CTLU LCTL(KC_U)
-#define PREFIX LCTL(KC_K)
-#define CTLP LCTL(KC_P)
-#define CTLR LCTL(KC_R)
-#define CTLL LCTL(KC_L)
-#define CTLE LCTL(KC_E)
-#define CTLN LCTL(KC_N)
-#define CTLM LCTL(KC_M)
-#define CTLU LCTL(KC_U)
-#define CTLY LCTL(KC_Y)
-#define CTLO LCTL(KC_O)
-#define CTLJ LCTL(KC_J)
-#define CTLG LCTL(KC_G)
-#define CTLH LCTL(KC_H)
-#define CTLTAB LCTL(KC_TAB)
-#define CTLENT LCTL(KC_ENT)
-#define BUILD LCTL(KC_B)
-#define SWITCH LALT(KC_O)
-#define PGDN_CTL RCTL_T(KC_PGDN)
-#define PGUP_SYM LT(_SYM_NUM, KC_PGUP)
-#define GRV_GUI RGUI_T(KC_GRV)
-#define LBRC_ALT LALT_T(KC_LBRC)
-#define RBRC_CTL LCTL_T(KC_RBRC)
-#define LBRC_GUI LGUI_T(KC_LBRC)
-#define RBRC_ALT LALT_T(KC_RBRC)
-#define LBRC_AGR RALT_T(KC_LBRC)
-#define SCLN_CTL LCTL_T(KC_SCLN)
-#define QUOT_AGR RALT_T(KC_QUOT)
-#define MINS_AGR RALT_T(KC_MINS)
-#define ONE_CTL RCTL_T(KC_1)
-#define TWO_AGR RALT_T(KC_2)
-#define THRE_ALT LALT_T(KC_3)
-#define EQL_GUI RGUI_T(KC_EQL)
 
+enum layer_names {
+    _QWERTY,
+    _NAVI_FN,
+    _SYM_NUM
+};
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-  [_ALPHA] = LAYOUT(
-    XXXXXXX,  KC_Q,     KC_W,     KC_B,     KC_R,     KC_F,       KC_Y,     KC_U,     KC_I,     KC_O,     KC_P,     XXXXXXX,
-    KC_TAB,   KC_A,     KC_S,     KC_D,     KC_T,     G_AGR,      H_AGR,    KC_N,     KC_E,     KC_L,     KC_K,     KC_ENT,
-              Z_GUI,    X_ALT,    C_SYM,    V_CTL,                          M_CTL,    KC_COMM,  DOT_ALT,  J_GUI,
-              KC_MNXT,  LOCK_NAV, OSM_LSFT, LOCK_ALP, XXXXXXX,    XXXXXXX,  QK_GESC,  SPC_CHNG, KC_BSPC,  KC_MPLY,  DB_TOGG
+  [_QWERTY] = LAYOUT(
+    KC_B,    KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,       KC_H,    KC_U,    KC_I,    KC_O,    KC_P,    KC_Y,
+    KC_TAB,  KC_A,    KC_S,    KC_D,    KC_F,    KC_G,       KC_N,    KC_J,    KC_K,    KC_L,    KC_SCLN, KC_ENT,
+             KC_Z,    KC_X,    KC_C,    KC_V,                         KC_M,    KC_COMM, KC_DOT,  KC_QUOT,
+             KC_MNXT, OSM_ALT, BS_SYM,  ESC_CTL, OSM_SFT,    OSM_AGR, DEL_GUI, SPC_NAV, OSM_ALT, KC_MPLY, DB_TOGG
   ),
-  [_NAVI] = LAYOUT(
-    XXXXXXX,  CTLY,     XXXXXXX,  KC_ACL0,  KC_ACL1,  KC_BTN1,    KC_HOME,  KC_WH_D,  KC_WH_U,  XXXXXXX,  XXXXXXX,  XXXXXXX,
-    _______,  UNDO,     CUT,      COPY,     PASTE,    KC_BTN2,    KC_LEFT,  KC_DOWN,  KC_UP,    KC_RGHT,  KC_END,   _______,
-              KC_LGUI,  KC_LALT,  MO_SYM,   KC_LCTL,                        PGDN_CTL, KC_PGUP,  KC_LALT,  KC_LGUI,
-              _______,  _______,  _______,  _______,  XXXXXXX,    XXXXXXX,  _______,  _______,  _______,  _______,  EE_CLR
+  [_NAVI_FN] = LAYOUT(
+    KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,      KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,
+    KC_TAB,  KC_HOME, KC_PGUP, KC_PGDN, KC_END,  KC_TILD,    KC_GRV,  KC_LEFT, KC_UP,   KC_RGHT, KC_WREF, KC_ENT,
+             KC_UNDO, KC_CUT,  KC_COPY, KC_PSTE,                      KC_BTN1, KC_DOWN, KC_BTN2, KC_WBAK,
+             KC_MPRV, _______, BS_GUI, _______, _______,     _______, _______, _______, _______, KC_PWR,  _______
   ),
   [_SYM_NUM] = LAYOUT(
-    XXXXXXX,  XXXXXXX,  XXXXXXX,  XXXXXXX,  XXXXXXX,  XXXXXXX,    KC_BSLS,  KC_7,     KC_8,     KC_9,     KC_MINS,  XXXXXXX,
-    KC_GRV,   XXXXXXX,  XXXXXXX,  XXXXXXX,  KC_DOT,   OSM_AGR,    KC_QUOT,  KC_4,     KC_5,     KC_6,     KC_EQL,   KC_SCLN,
-              GRV_GUI,  LBRC_ALT, _______,  RBRC_CTL,                       ONE_CTL,  KC_2,     THRE_ALT, SLSH_GUI,
-              KC_VOLD,  KC_LCBR,  _______,  KC_RCBR,  XXXXXXX,    XXXXXXX,  KC_COMM,  _______,  KC_0,     KC_VOLU,  QK_BOOT
-  ),
-  [_FUNC] = LAYOUT(
-    XXXXXXX,  QUIT,     CLOSE,    BUILD,    CTLR,     FIND,       CTLY,     KC_F7,    KC_F8,    KC_F9,    KC_F11,   XXXXXXX,
-    CTLTAB,   ALL,      SAVE,     MARK,     CTLE,     CTLG,       CTLH,     KC_F4,    KC_F5,    KC_F6,    KC_F12,   CTLENT,
-              KC_LGUI,  KC_LALT,  PREFIX,   KC_LCTL,                        KC_F1,    KC_F2,    KC_F3,    CTLJ,
-              KC_MPRV,  SWITCH,   OSM_LSFT, CTLO,     XXXXXXX,    XXXXXXX,  KC_PWR,   _______,  KC_F10,   KC_MNXT,  XXXXXXX
+    KC_EXLM, KC_AT,   KC_HASH, KC_DLR,  KC_PERC, KC_CIRC,    KC_AMPR, KC_7,    KC_8,    KC_9,    KC_ASTR, KC_PLUS,
+    KC_QUES, KC_SLSH, KC_LBRC, KC_LCBR, KC_LPRN, KC_PIPE,    KC_0,    KC_4,    KC_5,    KC_6,    KC_MINS, KC_EQL,
+             KC_BSLS, KC_RBRC, KC_RCBR, KC_RPRN,                      KC_1,    KC_2,    KC_3,    KC_UNDS,
+             KC_VOLD, _______, _______, _______, _______,    _______, _______, SPC_CTL, _______, KC_VOLU, _______
   )
 };
-
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  // Store the current modifier state in the variable for later reference.
-  static uint8_t mod_state;
-  mod_state = get_mods();
-  switch (keycode) {
-    case KC_BSPC:
-      {
-      // Keep track of whether the delete key status is registered.
-      static bool delkey_registered;
-      if (record->event.pressed) {
-        if (mod_state & MOD_MASK_SHIFT) {
-          // Make sure that Shift is not applied to the KC_DEL keycode.
-          del_mods(MOD_MASK_SHIFT);
-          register_code(KC_DEL);
-          // Update the boolean variable to reflect the status of KC_DEL
-          delkey_registered = true;
-          // Reapplying modifier state so that the held shift key(s)
-          // still work even after having tapped the Backspace/Delete key.
-          set_mods(mod_state);
-          return false;
-        }
-      } else { // on release of KC_BSPC
-        // In case KC_DEL is still being sent even after the release of KC_BSPC
-        if (delkey_registered) {
-          unregister_code(KC_DEL);
-          delkey_registered = false;
-          return false;
-        }
-      }
-    }
-  }
-  return true;
-}
